Implement static handling of the XDG_RUNTIME_DIR directory
In normal desktop situations XDG_RUNTIME_DIR would be provided when launching a user
session by the environment but as we're not in such a case (yet?) we have to take care
about this on our own. Sadly the site effect is that we have to tell every single launched
application about the correct path.

Signed-off-by: Simon Busch <morphis@gravedo.de>