Rework directory layout to represent the new shell concept
We're going to make luna-next captable of loading one specific shell on startup so the
user can switch between different implementations if he wants. Currently it's not possible
to load another shell but this will be implemented soon through a plugin mechanism.

The cardshell will be maintained within another repository.

Signed-off-by: Simon Busch <morphis@gravedo.de>