plugins: compositor: already add windows to model when created
We switch the way we handle windows with this a bit. When a window is now created it gets
directly added to the WindowModels. This gives the shell the ability to show a splash
screen until the window is fully mapped and we don't have to implement this on the client
side. To indicate the mapping status of a window there is now a property for this.

Signed-off-by: Simon Busch <morphis@gravedo.de>